Use DevLink component
